# docs/openapi.yaml
openapi: 3.0.3
info:
  title: AgentOps API
  version: "0.1.0"
  description: |
    REST API for the merged `agentops-server` process — composes routes from
    `agentops-api`, `docbrain-api` (nested under `/docbrain`), and
    `agentops-heavy-api` (accounts/auth, team management, repository
    connection/indexing, search, and webhooks) onto one Axum router, one
    port.

    Three auth models coexist depending on the route group:

    - **Session auth** (`sessionAuth`): a Bearer token from `POST /auth/login`
      or `POST /auth/signup`, used by the web UI and anything acting as a
      specific user.
    - **API key auth** (`apiKeyAuth`): a Bearer token from `agentops api-key
      generate` / `POST /auth/api-keys`, used by CLI/MCP/service callers.
    - **Webhook signature** (no Bearer token): `/webhooks/*` routes verify an
      HMAC signature header instead — see each route's description.

    Most `/repos/*` and `/search/*` routes accept *either* a session or an
    API key (`sessionOrApiKey`); the tenant resolves from the session if
    present, else from a `tenant`/`org` request field.
  license:
    name: MIT OR Apache-2.0
    url: https://github.com/Jesus-Glez60/agentops/blob/main/LICENSE-MIT
servers:
  - url: http://127.0.0.1:8420
    description: Default local bind address (AGENTOPS_ADDR) — change to whatever host/reverse-proxy this instance is actually deployed behind.
...
security:
  - {}
paths:
...
  /tools:
    get:
      tags: [Tools]
      summary: List agentops-api's MCP-style tool table
      security: [{ apiKeyAuth: [] }]
      responses:
        "200": { description: OK }
  /tools/{name}:
    post:
      tags: [Tools]
      summary: Call a tool by name (agentops-api's table)
      security: [{ apiKeyAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema: { type: object, description: "Tool-specific arguments — see GET /tools for each tool's input schema." }
      responses:
        "200": { description: Tool result }
  /docbrain/tools:
    get:
      tags: [Tools]
      summary: List docbrain-api's MCP-style tool table
      security: [{ apiKeyAuth: [] }]
      responses:
        "200": { description: OK }
  /docbrain/tools/{name}:
    post:
      tags: [Tools]
      summary: Call a docbrain tool by name (e.g. register_library, scrape_library, search_docs)
      security: [{ apiKeyAuth: [] }]
      parameters:
        - { name: name, in: path, required: true, schema: { type: string } }
      requestBody:
        content:
          application/json:
            schema: { type: object, description: "Tool-specific arguments — see GET /docbrain/tools for each tool's input schema." }
      responses:
        "200": { description: Tool result }

  # ---------------------------------------------------------------------
  # Libraries (docbrain)
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    sessionAuth:
      type: http
      scheme: bearer
      description: Session token from POST /auth/login or POST /auth/signup.
    apiKeyAuth:
      type: http
      scheme: bearer
      description: API key from `agentops api-key generate` or POST /auth/api-keys.
    sessionOrApiKey:
      type: http
      scheme: bearer
      description: Either a session token or an API key — tenant resolves from the session if present, else from a request field.
...